requirements.txt
apns2==0.7.1
appdirs==1.4.3
Babel==2.7.0
certifi==2019.9.11
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
distlib==0.3.0
Django==2.2.6
django-cors-headers==3.2.1
django-crontab==0.7.1
django-filter==2.2.0
django-phonenumber-field==3.0.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
et-xmlfile==1.0.1
fcm-django==0.3.4
filelock==3.0.12
gunicorn==20.1.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
jdcal==1.4.1
lml==0.0.9
openpyxl==3.0.4
phonenumbers==8.10.23
Pillow==6.2.1
psycopg2==2.8.4
py-vapid==1.7.1
pycparser==2.20
pyexcel==0.6.2
pyexcel-io==0.5.20
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.8
pyfcm==1.4.7
PyJWT==1.7.1
pytz==2019.3
pywebpush==1.11.0
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
texttable==1.6.2
urllib3==1.25.7
virtualenv==20.0.10
virtualenv-clone==0.5.3
whitenoise==5.2.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
